+++ src/main/java/kitchenpos/application/MenuService.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 import java.util.List;
 import kitchenpos.application.exception.MenuGroupNotFoundException;
 import kitchenpos.application.exception.ProductNotFoundException;
-import kitchenpos.repository.MenuRepository;
+import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.repository.MenuGroupRepository;
 import kitchenpos.repository.ProductRepository;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.Product;
 import kitchenpos.ui.dto.request.CreateMenuProductRequest;
 import kitchenpos.ui.dto.request.CreateMenuRequest;
