+++ src/main/java/kitchenpos/application/OrderService.java
@@ -6,10 +6,10 @@ import java.util.List;
 import kitchenpos.application.exception.MenuNotFoundException;
 import kitchenpos.application.exception.OrderNotFoundException;
 import kitchenpos.application.exception.OrderTableNotFoundException;
-import kitchenpos.repository.MenuRepository;
+import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
